+++ extension_packages/figpack_spike_sorting/src/views/component-time-scroll-view-3/useTimeScrollMouseWithModes.ts
@@ -1,6 +1,6 @@
 import { useCallback, useRef, useState } from "react";
 import { InteractionMode } from "./TimeScrollToolbar";
-import { useTimeRange } from "../../TimeseriesSelectionContext";
+import { useTimeRange } from "../context-timeseries-selection/TimeseriesSelectionContext";
 
 type MouseData = {
   mouseDownAchorX: number | null;