+++ extension_packages/figpack_spike_sorting/src/views/context-timeseries-selection/index.ts
@@ -1,0 +1,5 @@
+export { ProvideTimeseriesSelection } from "./ProvideTimeseriesSelection";
+export {
+  useTimeRange,
+  useTimeseriesSelection,
+} from "./TimeseriesSelectionContext";