+++ extension_packages/figpack_spike_sorting/src/views/view-spike-amplitudes/SpikeAmplitudesView.tsx
@@ -1,4 +1,4 @@
-import { useTimeseriesSelection } from "../../TimeseriesSelectionContext";
+import { useTimeseriesSelection } from "../context-timeseries-selection/TimeseriesSelectionContext";
 import { FunctionComponent, useEffect, useMemo, useState } from "react";
 import {
   idToNum,