+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+- Improvements to TimeseriesGraph for interval series data
+
 ## [0.3.2] - 2025-11-05
 
 - Create figpack-serve with worker for preparing and serving figpack figures from source URLs